deploy-on-ec2/backend/requirements.txt
fastapi[standard]
uvicorn
sqlalchemy
python-dotenv
psycopg2-binary
